feat: implement initial version of vue-accessible-color-picker
Implements a feature-rich, accessible color picker component for Vue.js. It comes with a color space representation in the form of an HSV cylinder’s hue slice. You can control a color’s saturation and lightness by moving around the thumb on the color space. The hue and alpha channel can be controlled via two range inputs. The numerical values that make up the selected color are shown in text fields. The text fields are shown based on what’s selected as the active format. Currently, four formats are available: HEX, HSL, HSV, and RGB. Finally, the selected color can be copied. The copied color will be a valid CSS color string the active format.

The color picker is available in two variants: Fully styled and almost fully unstyled. For both variants, a dedicated bundle can be found in the `dist` directory.