build(dependencies): install @semantic-release/github@7.0.6
Installs the latest version of @semantic-release/github due to bug #268 that caused the release workflow for vue-accessible-color-picker@1.0.0 to not complete properly. This should allow the next release to work properly. Once the semantic-release package is updated and ships with an updated version of the github plugin, this explicit dependency should be removed.