chore(husky): upgrade husky to v5
Upgrades Husky from version 4 to version 5. This requires a manual creation of git hooks in the `.husky/` directory rather than being able to specify the hooks in the `husky.config.js` file. This added verbosity is quite unfortunate, but I rather not have an unupgradable development dependency, so I did the upgrade anyway. Hopefully in the future, an abstraction like the `husky.config.js` file offered, will become available again.