feat: set different default format via prop
Introduces a new optional prop `defaultFormat` that when set causes the color picker to be rendered with the respective format. By default, the prop is set to `'rgb'`. It can be set to `'hex'`, `'hsl'`, `'hwb'`, and `'rgb'` (i.e. any allowed visible color format). In the future, the default color format might change to `'hsl'` as I consider it the generally more useful color format.