chore: adds rudimentary demo page
Adds a very basic demo page to the projects. This makes testing the behavior of the color picker easier which became necessary with Vue.js 3 effectively making it impossible to use `npm link` to test Vue plugins. That’s because such a scenario would typically lead to two separate Vue instances being used which leads to conflicts with Vue’s use of `Symbol` which will be from two different script realms.